refactor(inf-146): deepen encounter transition seam
