+++ openspec/changes/inf-146-architecture-1-deepen-encounter-transition-module/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-11
